Added UrlResolver class
